+++ carp_core/lib/carp_client/data_manager.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Copenhagen Center for Health Technology (CACHET) at the
+ * Copyright 2021-2023 Copenhagen Center for Health Technology (CACHET) at the
  * Technical University of Denmark (DTU).
  * Use of this source code is governed by a MIT-style license that can be
  * found in the LICENSE file.
@@ -7,33 +7,39 @@
 
 part of carp_core_client;
 
-/// The [DataManager] interface is used to upload [DataPoint] objects to any
+/// The [DataManager] interface is used to upload [DataPoint] data to any
 /// data manager that implements this interface.
+///
+/// Note that each instance of a data manager supports one deployment
+/// ([MasterDeviceDeployment]). A data manager should hence be able to handle
+/// separate data management of concurrently running deployments.
+/// Hence, caution on resource starvation should be considered, such as not
+/// accessing the same file or network socket.
 abstract class DataManager {
-  /// The deployment using this data manager
+  /// The deployment using this data manager.
   MasterDeviceDeployment get deployment;
-
-  /// The ID of the study deployment that this manager is handling.
-  String get studyDeploymentId;
 
   /// The type of this data manager as enumerated in [DataEndPointTypes].
   String get type;
 
-  /// Initialize the data manager by specifying the study [deployment], the
-  /// [dataEndPoint], and the stream of [data] events to handle.
+  /// Initialize the data manager by specifying the [dataEndPoint], study
+  /// [deployment], and the stream of [data] events to handle.
   Future<void> initialize(
-    MasterDeviceDeployment deployment,
     DataEndPoint dataEndPoint,
+    MasterDeviceDeployment deployment,
     Stream<DataPoint> data,
   );
 
-  /// Close the data manager (e.g. closing connections).
+  /// Flush any buffered data and close this data manager.
+  /// After calling [close] the data manager can no longer be used.
   Future<void> close();
 
   /// Stream of data manager events.
   Stream<DataManagerEvent> get events;
 
-  /// On each data event from the data stream, the [onDataPoint] handler is called.
+  /// On each data event from the data stream, the [onDataPoint] handler is
+  /// called. Implementations of this interface should handle how to save
+  /// or upload the [dataPoint].
   Future<void> onDataPoint(DataPoint dataPoint);
 
   /// When the data stream closes, the [onDone] handler is called.
@@ -55,11 +61,8 @@ class DataManagerEvent {
   String toString() => 'DataManagerEvent - type: $type';
 }
 
-/// An enumeration of data manager event types
+/// An enumeration of data manager event types.
 class DataManagerEventTypes {
-  /// DATA MANAGER INITIALIZED event
-  static const String INITIALIZED = 'initialized';
-
-  /// DATA MANAGER CLOSED event
-  static const String CLOSED = 'closed';
+  static const String INITIALIZED = 'INITIALIZED';
+  static const String CLOSED = 'CLOSED';
 }