+++ carp_mobile_sensing/lib/runtime/study_controller.dart
@@ -39,9 +39,6 @@ class SmartphoneDeploymentController extends StudyRuntime {
   /// The datum transformed used to transform data before upload.
   DatumTransformer get transformer => _transformer;
 
-  /// The permissions granted to this study from the OS.
-  Map<Permission, PermissionStatus>? permissions;
-
   /// The stream of all sampled data points.
   ///
   /// Data points in the [data] stream are transformed in the following order:
@@ -70,8 +67,8 @@ class SmartphoneDeploymentController extends StudyRuntime {
   /// objects that has been collected.
   int get samplingSize => _samplingSize;
 
-  /// Create a new [SmartphoneDeploymentController] to control the runtime behavior
-  /// of a study deployment.
+  /// Create a new [SmartphoneDeploymentController] to control the runtime
+  /// behavior of a study deployment.
   SmartphoneDeploymentController(super.deploymentService, super.deviceRegistry);
 
   /// Verifies whether the master device is ready for deployment and in case
@@ -155,14 +152,16 @@ class SmartphoneDeploymentController extends StudyRuntime {
   }
 
   /// Erase all study deployment information cached locally on this phone.
+  /// Note that all data sampled and cached locally on the phone is also deleted.
+  ///
   /// Returns `true` if successful.
   Future<bool> eraseDeployment() async {
     bool success = true;
 
     try {
-      String name = (await filename)!;
-      info("Erasing deployment cache from file '$name'.");
-      await File(name).delete();
+      String name = await Settings().getDeploymentBasePath(studyDeploymentId!);
+      info("Erasing deployment cache for deployment '$studyDeploymentId'.");
+      await File(name).delete(recursive: true);
     } catch (exception) {
       success = false;
       warning('Failed to delete deployment - $exception');
@@ -191,18 +190,11 @@ class SmartphoneDeploymentController extends StudyRuntime {
   ///    * [transformer] - a generic [DatumTransformer] function which transform
   ///      each collected [Datum]. If not specified, a 1:1 mapping is done,
   ///      i.e. no transformation.
-  ///    * [askForPermissions] - automatically ask for permissions for all sampling
-  ///      packages at once. Default to `true`. If you want the app to handle
-  ///      permissions, set this to `false`.
-  ///    * [enableNotifications] - should notification be enabled and send to the user
-  ///      when an app task is triggered?
   ///
   Future<void> configure({
     DataEndPoint? dataEndPoint,
     String privacySchemaName = NameSpace.CARP,
     DatumTransformer? transformer,
-    bool askForPermissions = true,
-    bool enableNotifications = true,
   }) async {
     assert(deployment != null,
         'Cannot configure a StudyDeploymentController without a deployment.');
@@ -213,10 +205,6 @@ class SmartphoneDeploymentController extends StudyRuntime {
     // initialize all devices from the master deployment, incl. this master device
     deviceRegistry.initializeDevices(deployment!);
 
-    // initialize the app task controller singleton
-    await AppTaskController()
-        .initialize(enableNotifications: enableNotifications);
-
     _executor = StudyDeploymentExecutor();
 
     // initialize optional parameters
@@ -225,7 +213,7 @@ class SmartphoneDeploymentController extends StudyRuntime {
     _transformer = transformer ?? ((datum) => datum);
 
     if (_dataEndPoint != null) {
-      _dataManager = DataManagerRegistry().lookup(_dataEndPoint!.type);
+      _dataManager = DataManagerRegistry().create(_dataEndPoint!.type);
     }
 
     if (_dataManager == null) {
@@ -236,13 +224,10 @@ class SmartphoneDeploymentController extends StudyRuntime {
     // if no user is specified for this study, look up the local user id
     deployment!.userId ??= await Settings().userId;
 
-    // setting up permissions
-    if (askForPermissions) await askForAllPermissions();
-
     // initialize the data manager, device registry, and study executor
     await _dataManager?.initialize(
-      deployment!,
       deployment!.dataEndPoint!,
+      deployment!,
       data,
     );
 
@@ -263,8 +248,8 @@ class SmartphoneDeploymentController extends StudyRuntime {
     print('       user id : ${deployment!.userId}');
     print('      platform : ${DeviceInfo().platform.toString()}');
     print('     device ID : ${DeviceInfo().deviceID.toString()}');
-    print('  data manager : $_dataManager');
     print(' data endpoint : $_dataEndPoint');
+    print('  data manager : $_dataManager');
     print('        status : ${status.toString().split('.').last}');
     print('===============================================================');
   }
@@ -299,22 +284,6 @@ class SmartphoneDeploymentController extends StudyRuntime {
   // /// Disable power-aware sensing.
   // void disablePowerAwareness() => _battery.stop();
 
-  /// Asking for all permissions needed for the included sampling packages.
-  ///
-  /// Should be called before sensing is started, if not already done as part of
-  /// [configure].
-  Future<void> askForAllPermissions() async {
-    if (SamplingPackageRegistry().permissions.isNotEmpty) {
-      info('Asking for permission for all measure types.');
-      permissions = await SamplingPackageRegistry().permissions.request();
-
-      for (var permission in SamplingPackageRegistry().permissions) {
-        PermissionStatus status = await permission.status;
-        info('Permissions for $permission : $status');
-      }
-    }
-  }
-
   /// Start this controller and if [resume] is true, resume data collection
   /// according to the parameters specified in [configure].
   ///
@@ -326,41 +295,32 @@ class SmartphoneDeploymentController extends StudyRuntime {
         '$runtimeType - Cannot resume this controller, since the the runtime is not initialized. '
         'Call the configure() method first.');
 
-    info('$runtimeType - Starting data sampling ...');
+    info('$runtimeType - Starting data sampling...');
     super.start();
     if (resume) executor!.resume();
   }
 
   @override
   void pause() {
-    info('$runtimeType - Pausing data sampling ...');
+    info('$runtimeType - Pausing data sampling...');
     super.pause();
     executor!.pause();
   }
 
   @override
-  void dispose() {
-    info('$runtimeType - Disposing ...');
-    executor?.stop();
-    dataManager?.close().then((_) => super.dispose());
-  }
-
-  @override
   Future<void> remove() async {
-    info('$runtimeType - Removing deployment from this smartphone ...');
-    dispose();
-    AppTaskController().removeStudyDeployment(studyDeploymentId!);
-    await AppTaskController().saveQueue();
+    info('$runtimeType - Removing deployment from this smartphone...');
+    executor?.stop();
+    await dataManager?.close();
     await eraseDeployment();
     await super.remove();
   }
 
   @override
-  Future<void> stop() async {
-    info(
-        '$runtimeType - Permanently stopping data sampling on this smartphone...');
-    await remove();
-    super.stop();
+  void dispose() {
+    info('$runtimeType - Disposing deployment...');
+    executor?.stop();
+    dataManager?.close().then((_) => super.dispose());
   }
 }