+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.40.12
 
+* Refactor of `DataManager` interfaces; file, sqlite, etc. to support multiple deployments on the same client.
 * NotificationController now cancels notifications for a study which is removed.
 
 ## 0.40.11
@@ -8,9 +9,9 @@
 
 ## 0.40.10
 
-* adding `dispose`, `pause`, and `remove` life cycle methods to the `DeploymentController`
-* support for deleting cached runtime information (deployment and task list)
-* fix of bug in `getTaskExecutor` method
+* Adding `dispose`, `pause`, and `remove` life cycle methods to the `DeploymentController`
+* Support for deleting cached runtime information (deployment and task list)
+* Fix of bug in `getTaskExecutor` method
 
 ## 0.40.7